+++ Twist/src/com/secondhand/model/PowerUp.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.math.Vector2;
 
 public abstract class PowerUp extends Entity {
 	
-	public PowerUp (Vector2 position, int radius) {
+	public PowerUp (Vector2 position, float radius) {
 		super(position,radius);
 	}
 }
